// Cesium/src/Wes.d.ts
interface WesDatasources extends DataSourceCollection {
    _dataSources: WesDataSource[];
}
// ...
type CesiumWindow = Window &
    typeof globalThis & {
        sourcesWithLegends: Accessor<any[]>;
        setSourcesWithLegends: Setter<any[]>;
        setTimeMap: Setter<Map<number, number>>;
        timeMap: Accessor<Map<number, number>>;
        Map3DViewer: Viewer;
        Map3DController: Map3DController;
        optionsMap: Accessor<any>;
    };
// ...
type Map3DController = any;
// ...
    export declare class WesDataSource extends DataSource {
        Constructor: WesDataSource;
        _name: string;
        _changed: CesiumEvent;
        _error: CesiumEvent;
        _isLoading: boolean;
        _loading: CesiumEvent;
        _isLoaded: boolean;
        _type: string;
        _entityCollection: EntityCollection;
        _entityCluster: EntityCluster;
        _url: string;
        _show: boolean;
        _viewer: Viewer;
        _scratchRectangle: Rectangle;
        _isInitialized: boolean;
        _isCancelledIdMap: Set<number>;
        _currentLoadId: number;
        _fetchRequestAbortController: AbortController;
        _removed: boolean;
        initialize: Function;
        stop: Function;
        isCancelled: Function;
        _setLoading: Function;
        fetchJson: Function;
    }